+++ app/club.php
@@ -5,12 +5,17 @@ namespace App;
 use Illuminate\Database\Eloquent\Model;
 use Auth;
 use DB;
+use Illuminate\Database\Eloquent\SoftDeletes;
 class club extends Model
 {
-
+    use SoftDeletes;
     public function user()
     {
         return $this->belongsTo(User::class);
+    }
+    public function lastAdminEdit()
+    {
+        return $this->belongsTo(User::class,'last_admin_edit');
     }
     public function comments()
     {
@@ -59,17 +64,23 @@ class club extends Model
         parent::boot();
         static::deleting(function($club) {
             if($club->club_price_file!=''){
-                $path_to_file = explode('storage/',$club->club_price_file)[1];
-                if(file_exists(storage_path('app/public/'.$path_to_file))){
-                unlink(storage_path('app/public/'.$path_to_file));}
+                $path_to_file = explode('storage/',$club->club_price_file);
+                if(isset($path_to_file[1])){
+                    $path_to_file = $path_to_file[1];
+                    if(file_exists(storage_path('app/public/'.$path_to_file))){
+                        unlink(storage_path('app/public/'.$path_to_file));}
+                }
              }
              if($club->club_photos!=''){
                  $images = explode(',',$club->club_photos);
                 foreach($images as $link){
-                    $path_to_file = explode('storage/',$link)[1];
-                    if(file_exists(storage_path('app/public/'.$path_to_file))){
-                        unlink(storage_path('app/public/'.$path_to_file));
-                    }
+                    $path_to_file = explode('storage/',$link);
+                        if(isset($path_to_file[1])){
+                            $path_to_file = $path_to_file[1];
+                            if(file_exists(storage_path('app/public/'.$path_to_file))){
+                                unlink(storage_path('app/public/'.$path_to_file));
+                            }
+                        }
                 }
              }
         }
